requirements.txt
flask
python-binance
cryptography
aiohttp
requests